Remove legacy Kotlin sync client
