+++ common/src/commonMain/kotlin/com/powersync/bucket/BucketStorageImpl.kt
@@ -9,11 +9,7 @@ import com.powersync.db.internal.InternalDatabase
 import com.powersync.db.internal.InternalTable
 import com.powersync.db.internal.PowerSyncTransaction
 import com.powersync.sync.Instruction
-import com.powersync.sync.LegacySyncImplementation
-import com.powersync.sync.SyncDataBatch
-import com.powersync.sync.SyncLocalDatabaseResult
 import com.powersync.utils.JsonUtil
-import kotlinx.serialization.Serializable
 
 internal class BucketStorageImpl(
     private val db: InternalDatabase,
@@ -113,65 +109,6 @@ internal class BucketStorageImpl(
         }
     }
 
-    @LegacySyncImplementation
-    override suspend fun saveSyncData(syncDataBatch: SyncDataBatch) {
-        db.writeTransaction { tx ->
-            val jsonString = JsonUtil.json.encodeToString(syncDataBatch)
-            tx.execute(
-                "INSERT INTO powersync_operations(op, data) VALUES(?, ?)",
-                listOf("save", jsonString),
-            )
-        }
-    }
-
-    @LegacySyncImplementation
-    override suspend fun getBucketStates(): List<BucketState> =
-        db.getAll(
-            "SELECT name AS bucket, CAST(last_op AS TEXT) AS op_id FROM ${InternalTable.BUCKETS} WHERE pending_delete = 0 AND name != '\$local'",
-            mapper = { cursor ->
-                BucketState(
-                    bucket = cursor.getString(0)!!,
-                    opId = cursor.getString(1)!!,
-                )
-            },
-        )
-
-    @LegacySyncImplementation
-    override suspend fun getBucketOperationProgress(): Map<String, LocalOperationCounters> =
-        buildMap {
-            val rows =
-                db.getAll("SELECT name, count_at_last, count_since_last FROM ps_buckets") { cursor ->
-                    cursor.getString(0)!! to
-                        LocalOperationCounters(
-                            atLast = cursor.getLong(1)!!.toInt(),
-                            sinceLast = cursor.getLong(2)!!.toInt(),
-                        )
-                }
-
-            for ((name, counters) in rows) {
-                put(name, counters)
-            }
-        }
-
-    @LegacySyncImplementation
-    private suspend fun deleteBucket(bucketName: String) {
-        db.writeTransaction { tx ->
-            tx.execute(
-                "INSERT INTO powersync_operations(op, data) VALUES(?, ?)",
-                listOf("delete_bucket", bucketName),
-            )
-        }
-
-        Logger.d("[deleteBucket] Done deleting")
-    }
-
-    @LegacySyncImplementation
-    override suspend fun removeBuckets(bucketsToDelete: List<String>) {
-        bucketsToDelete.forEach { bucketName ->
-            deleteBucket(bucketName)
-        }
-    }
-
     override suspend fun hasCompletedSync(): Boolean {
         if (hasCompletedSync.value) {
             return true
@@ -191,160 +128,6 @@ internal class BucketStorageImpl(
         } else {
             false
         }
-    }
-
-    @LegacySyncImplementation
-    override suspend fun syncLocalDatabase(
-        targetCheckpoint: Checkpoint,
-        partialPriority: StreamPriority?,
-    ): SyncLocalDatabaseResult {
-        val result = validateChecksums(targetCheckpoint, partialPriority)
-
-        if (!result.checkpointValid) {
-            logger.w { "[SyncLocalDatabase] Checksums failed for ${result.checkpointFailures}" }
-            result.checkpointFailures?.forEach { bucketName ->
-                deleteBucket(bucketName)
-            }
-            result.ready = false
-            return result
-        }
-
-        val bucketNames =
-            targetCheckpoint.checksums
-                .let {
-                    if (partialPriority == null) {
-                        it
-                    } else {
-                        it.filter { cs -> cs.priority >= partialPriority }
-                    }
-                }.map { it.bucket }
-
-        db.writeTransaction { tx ->
-            tx.execute(
-                "UPDATE ps_buckets SET last_op = ? WHERE name IN (SELECT json_each.value FROM json_each(?))",
-                listOf(targetCheckpoint.lastOpId, JsonUtil.json.encodeToString(bucketNames)),
-            )
-
-            if (partialPriority == null && targetCheckpoint.writeCheckpoint != null) {
-                tx.execute(
-                    "UPDATE ps_buckets SET last_op = ? WHERE name = '\$local'",
-                    listOf(targetCheckpoint.writeCheckpoint),
-                )
-            }
-        }
-
-        val valid = updateObjectsFromBuckets(targetCheckpoint, partialPriority)
-
-        if (!valid) {
-            return SyncLocalDatabaseResult(
-                ready = false,
-                checkpointValid = true,
-            )
-        }
-
-        return SyncLocalDatabaseResult(
-            ready = true,
-        )
-    }
-
-    @LegacySyncImplementation
-    private suspend fun validateChecksums(
-        checkpoint: Checkpoint,
-        priority: StreamPriority? = null,
-    ): SyncLocalDatabaseResult {
-        val serializedCheckpoint =
-            JsonUtil.json.encodeToString(
-                when (priority) {
-                    null -> checkpoint
-                    // Only validate buckets with a priority included in this partial sync.
-                    else -> checkpoint.copy(checksums = checkpoint.checksums.filter { it.priority >= priority })
-                },
-            )
-
-        val res =
-            db.getOptional(
-                "SELECT powersync_validate_checkpoint(?) AS result",
-                parameters = listOf(serializedCheckpoint),
-                mapper = { cursor ->
-                    cursor.getString(0)!!
-                },
-            )
-                ?: // no result
-                return SyncLocalDatabaseResult(
-                    ready = false,
-                    checkpointValid = false,
-                )
-
-        return JsonUtil.json.decodeFromString<SyncLocalDatabaseResult>(res)
-    }
-
-    /**
-     * Atomically update the local state.
-     *
-     * This includes creating new tables, dropping old tables, and copying data over from the oplog.
-     */
-    @LegacySyncImplementation
-    private suspend fun updateObjectsFromBuckets(
-        checkpoint: Checkpoint,
-        priority: StreamPriority? = null,
-    ): Boolean {
-        @Serializable
-        data class SyncLocalArgs(
-            val priority: StreamPriority,
-            val buckets: List<String>,
-        )
-
-        val args =
-            if (priority != null) {
-                JsonUtil.json.encodeToString(
-                    SyncLocalArgs(
-                        priority = priority,
-                        buckets = checkpoint.checksums.filter { it.priority >= priority }.map { it.bucket },
-                    ),
-                )
-            } else {
-                ""
-            }
-
-        return db.writeTransaction { tx ->
-            tx.execute(
-                "INSERT INTO powersync_operations(op, data) VALUES(?, ?)",
-                listOf("sync_local", args),
-            )
-
-            val res =
-                tx.get("select last_insert_rowid()") { cursor ->
-                    cursor.getLong(0)!!
-                }
-
-            val didApply = res == 1L
-            if (didApply && priority == null) {
-                // Reset progress counters. We only do this for a complete sync, as we want a download progress to
-                // always cover a complete checkpoint instead of resetting for partial completions.
-                tx.execute(
-                    """
-                    UPDATE ps_buckets SET count_since_last = 0, count_at_last = ?1->name
-                      WHERE name != '${'$'}local' AND ?1->name IS NOT NULL
-                    """.trimIndent(),
-                    listOf(
-                        JsonUtil.json.encodeToString(
-                            buildMap<String, Int> {
-                                for (bucket in checkpoint.checksums) {
-                                    bucket.count?.let { put(bucket.bucket, it) }
-                                }
-                            },
-                        ),
-                    ),
-                )
-            }
-
-            return@writeTransaction didApply
-        }
-    }
-
-    @LegacySyncImplementation
-    override fun setTargetCheckpoint(checkpoint: Checkpoint) {
-        // No-op
     }
 
     private fun handleControlResult(cursor: SqlCursor): List<Instruction> {
