--- common/src/commonMain/kotlin/com/powersync/bucket/LocalOperationCounters.kt
@@ -1,9 +1,0 @@
-package com.powersync.bucket
-
-import com.powersync.sync.LegacySyncImplementation
-
-@LegacySyncImplementation
-internal data class LocalOperationCounters(
-    val atLast: Int,
-    val sinceLast: Int,
-)
