--- common/src/commonMain/kotlin/com/powersync/bucket/SqliteOp.kt
@@ -1,6 +1,0 @@
-package com.powersync.bucket
-
-internal data class SqliteOp(
-    val sql: String,
-    val args: List<Any>,
-)
