+++ common/src/commonMain/kotlin/com/powersync/sync/StreamingSync.kt
@@ -5,10 +5,7 @@ import co.touchlab.kermit.Severity
 import co.touchlab.stately.concurrency.AtomicReference
 import com.powersync.ExperimentalPowerSyncAPI
 import com.powersync.PowerSyncException
-import com.powersync.bucket.BucketChecksum
-import com.powersync.bucket.BucketRequest
 import com.powersync.bucket.BucketStorage
-import com.powersync.bucket.Checkpoint
 import com.powersync.bucket.PowerSyncControlArguments
 import com.powersync.bucket.WriteCheckpointResponse
 import com.powersync.connectors.PowerSyncBackendConnector
@@ -307,16 +304,6 @@ internal class StreamingSyncClient(
         return originalFlow.buffer(Channel.RENDEZVOUS)
     }
 
-    private fun receiveTextLines(req: JsonElement): Flow<String> =
-        syncEndpointFlow(req, supportBson = false) { isBson, response ->
-            status.update { copy(connected = true, connecting = false) }
-            check(!isBson)
-
-            response.body<ByteReadChannel>().lines().collect {
-                emit(it)
-            }
-        }
-
     private fun receiveTextOrBinaryLines(req: JsonElement): Flow<PowerSyncControlArguments> {
         val needsRSocket = httpClient.attributes[WebSocketIfNecessaryPlugin.needsRSocketKey]
 
@@ -353,28 +340,18 @@ internal class StreamingSyncClient(
 
     private suspend fun streamingSyncIteration(): SyncIterationResult =
         coroutineScope {
-            if (options.newClientImplementation) {
-                val iteration = ActiveIteration(this)
+            val iteration = ActiveIteration(this)
 
-                try {
-                    iteration.start()
-                } finally {
-                    // This can't be cancelled because we need to send a stop message, which is async, to
-                    // clean up resources.
-                    withContext(NonCancellable) {
-                        iteration.stop()
-                    }
+            try {
+                iteration.start()
+            } finally {
+                // This can't be canceled because we need to send a stop message, which is async, to
+                // clean up resources.
+                withContext(NonCancellable) {
+                    iteration.stop()
                 }
-            } else {
-                legacySyncIteration()
-                SyncIterationResult()
             }
         }
-
-    @OptIn(LegacySyncImplementation::class)
-    private suspend fun CoroutineScope.legacySyncIteration() {
-        LegacyIteration(this).streamingSyncIteration()
-    }
 
     /**
      * Implementation of a sync iteration that delegates to helper functions implemented in the
@@ -545,295 +522,6 @@ internal class StreamingSyncClient(
         }
     }
 
-    @LegacySyncImplementation
-    private inner class LegacyIteration(
-        val scope: CoroutineScope,
-    ) {
-        suspend fun streamingSyncIteration() {
-            check(schema.rawTables.isEmpty()) {
-                "Raw tables are only supported by the Rust sync client."
-            }
-
-            val bucketEntries = bucketStorage.getBucketStates()
-            val initialBuckets = mutableMapOf<String, String>()
-
-            var state =
-                SyncStreamState(
-                    targetCheckpoint = null,
-                    validatedCheckpoint = null,
-                    appliedCheckpoint = null,
-                    bucketSet = initialBuckets.keys.toMutableSet(),
-                )
-
-            bucketEntries.forEach { entry ->
-                initialBuckets[entry.bucket] = entry.opId
-            }
-
-            val req =
-                StreamingSyncRequest(
-                    buckets =
-                        initialBuckets.map { (bucket, after) ->
-                            BucketRequest(
-                                bucket,
-                                after,
-                            )
-                        },
-                    clientId = clientId!!,
-                    parameters = params,
-                    appMetadata = appMetadata,
-                )
-
-            lateinit var receiveLines: Job
-            receiveLines =
-                scope.launch {
-                    var hadLine = false
-                    receiveTextLines(JsonUtil.json.encodeToJsonElement(req)).collect { value ->
-                        val line = JsonUtil.json.decodeFromString<SyncLine>(value)
-
-                        if (!hadLine) {
-                            // Trigger a crud upload when receiving the first sync line: We could have
-                            // pending local writes made while disconnected, so in addition to listening on
-                            // updates to `ps_crud`, we also need to trigger a CRUD upload in some other
-                            // cases. We do this on the first sync line because the client is likely to be
-                            // online in that case.
-                            hadLine = true
-                            triggerCrudUploadAsync()
-                        }
-
-                        state = handleInstruction(line, value, state)
-
-                        if (state.abortIteration) {
-                            receiveLines.cancel()
-                        }
-                    }
-                }
-
-            receiveLines.join()
-            status.update { abortedDownload() }
-        }
-
-        private suspend fun handleInstruction(
-            line: SyncLine,
-            jsonString: String,
-            state: SyncStreamState,
-        ): SyncStreamState =
-            when (line) {
-                is SyncLine.FullCheckpoint -> {
-                    handleStreamingSyncCheckpoint(line, state)
-                }
-
-                is SyncLine.CheckpointDiff -> {
-                    handleStreamingSyncCheckpointDiff(line, state)
-                }
-
-                is SyncLine.CheckpointComplete -> {
-                    handleStreamingSyncCheckpointComplete(state)
-                }
-
-                is SyncLine.CheckpointPartiallyComplete -> {
-                    handleStreamingSyncCheckpointPartiallyComplete(
-                        line,
-                        state,
-                    )
-                }
-
-                is SyncLine.KeepAlive -> {
-                    handleStreamingKeepAlive(line, state)
-                }
-
-                is SyncLine.SyncDataBucket -> {
-                    handleStreamingSyncData(line, state)
-                }
-
-                SyncLine.UnknownSyncLine -> {
-                    logger.w { "Unhandled instruction $jsonString" }
-                    state
-                }
-            }
-
-        private suspend fun handleStreamingSyncCheckpoint(
-            line: SyncLine.FullCheckpoint,
-            state: SyncStreamState,
-        ): SyncStreamState {
-            val (checkpoint) = line
-            state.targetCheckpoint = checkpoint
-
-            val bucketsToDelete = state.bucketSet!!.toMutableList()
-            val newBuckets = mutableSetOf<String>()
-
-            checkpoint.checksums.forEach { checksum ->
-                run {
-                    newBuckets.add(checksum.bucket)
-                    bucketsToDelete.remove(checksum.bucket)
-                }
-            }
-
-            state.bucketSet = newBuckets
-            startTrackingCheckpoint(checkpoint, bucketsToDelete)
-
-            return state
-        }
-
-        private suspend fun startTrackingCheckpoint(
-            checkpoint: Checkpoint,
-            bucketsToDelete: List<String>,
-        ) {
-            val progress = bucketStorage.getBucketOperationProgress()
-            status.update {
-                copy(
-                    downloading = true,
-                    downloadProgress = SyncDownloadProgress(progress, checkpoint),
-                )
-            }
-
-            if (bucketsToDelete.isNotEmpty()) {
-                logger.i { "Removing buckets [${bucketsToDelete.joinToString(separator = ", ")}]" }
-            }
-
-            bucketStorage.removeBuckets(bucketsToDelete)
-            bucketStorage.setTargetCheckpoint(checkpoint)
-        }
-
-        private suspend fun handleStreamingSyncCheckpointComplete(state: SyncStreamState): SyncStreamState {
-            val checkpoint = state.targetCheckpoint!!
-            var result = bucketStorage.syncLocalDatabase(checkpoint)
-            val pending = isUploadingCrud.get()
-
-            if (!result.checkpointValid) {
-                // This means checksums failed. Start again with a new checkpoint.
-                // TODO: better back-off
-                delay(50)
-                state.abortIteration = true
-                // TODO handle retries
-                return state
-            } else if (!result.ready && pending != null) {
-                // We have pending entries in the local upload queue or are waiting to confirm a write checkpoint, which
-                // prevented this checkpoint from applying. Wait for that to complete and try again.
-                logger.d { "Could not apply checkpoint due to local data. Waiting for in-progress upload before retrying." }
-                pending.done.await()
-
-                result = bucketStorage.syncLocalDatabase(checkpoint)
-            }
-
-            if (result.checkpointValid && result.ready) {
-                state.appliedCheckpoint = checkpoint.clone()
-                logger.i { "validated checkpoint ${state.appliedCheckpoint}" }
-
-                state.validatedCheckpoint = state.targetCheckpoint
-                status.update { copyWithCompletedDownload() }
-            } else {
-                logger.d { "Could not apply checkpoint. Waiting for next sync complete line" }
-            }
-
-            return state
-        }
-
-        private suspend fun handleStreamingSyncCheckpointPartiallyComplete(
-            line: SyncLine.CheckpointPartiallyComplete,
-            state: SyncStreamState,
-        ): SyncStreamState {
-            val priority = line.priority
-            val result = bucketStorage.syncLocalDatabase(state.targetCheckpoint!!, priority)
-            if (!result.checkpointValid) {
-                // This means checksums failed. Start again with a new checkpoint.
-                // TODO: better back-off
-                delay(50)
-                state.abortIteration = true
-                // TODO handle retries
-                return state
-            } else if (!result.ready) {
-                // Checkpoint is valid, but we have local data preventing this to be published. We'll try to resolve this
-                // once we have a complete checkpoint if the problem persists.
-            } else {
-                logger.i { "validated partial checkpoint ${state.appliedCheckpoint} up to priority of $priority" }
-            }
-
-            status.update {
-                copy(
-                    priorityStatusEntries =
-                        buildList {
-                            // All states with a higher priority can be deleted since this partial sync includes them.
-                            addAll(status.priorityStatusEntries.filter { it.priority >= line.priority })
-                            add(
-                                PriorityStatusEntry(
-                                    priority = priority,
-                                    lastSyncedAt = Clock.System.now(),
-                                    hasSynced = true,
-                                ),
-                            )
-                        },
-                )
-            }
-            return state
-        }
-
-        private suspend fun handleStreamingSyncCheckpointDiff(
-            checkpointDiff: SyncLine.CheckpointDiff,
-            state: SyncStreamState,
-        ): SyncStreamState {
-            // TODO: It may be faster to just keep track of the diff, instead of the entire checkpoint
-            if (state.targetCheckpoint == null) {
-                throw Exception("Checkpoint diff without previous checkpoint")
-            }
-
-            val newBuckets = mutableMapOf<String, BucketChecksum>()
-
-            state.targetCheckpoint!!.checksums.forEach { checksum ->
-                newBuckets[checksum.bucket] = checksum
-            }
-            checkpointDiff.updatedBuckets.forEach { checksum ->
-                newBuckets[checksum.bucket] = checksum
-            }
-
-            checkpointDiff.removedBuckets.forEach { bucket -> newBuckets.remove(bucket) }
-
-            val newCheckpoint =
-                Checkpoint(
-                    lastOpId = checkpointDiff.lastOpId,
-                    checksums = newBuckets.values.toList(),
-                    writeCheckpoint = checkpointDiff.writeCheckpoint,
-                )
-
-            state.targetCheckpoint = newCheckpoint
-            startTrackingCheckpoint(newCheckpoint, checkpointDiff.removedBuckets)
-
-            return state
-        }
-
-        private suspend fun handleStreamingSyncData(
-            data: SyncLine.SyncDataBucket,
-            state: SyncStreamState,
-        ): SyncStreamState {
-            val batch = SyncDataBatch(listOf(data))
-            bucketStorage.saveSyncData(batch)
-            status.update {
-                copy(
-                    downloading = true,
-                    downloadProgress = downloadProgress?.incrementDownloaded(batch),
-                )
-            }
-            return state
-        }
-
-        private fun handleStreamingKeepAlive(
-            keepAlive: SyncLine.KeepAlive,
-            state: SyncStreamState,
-        ): SyncStreamState {
-            val (tokenExpiresIn) = keepAlive
-
-            if (tokenExpiresIn <= 0) {
-                // Connection would be closed automatically right after this
-                logger.i { "Token expiring reconnect" }
-                connector.invalidateCredentials()
-                state.abortIteration = true
-                return state
-            }
-            // Don't await the upload job, we can keep receiving sync lines
-            triggerCrudUploadAsync()
-            return state
-        }
-    }
-
     internal companion object Companion {
         // The sync service sends a token keepalive message roughly every 20 seconds. So if we don't receive a message
         // in twice that time, assume the connection is broken.
@@ -901,15 +589,6 @@ internal class StreamingSyncClient(
         }
     }
 }
-
-@LegacySyncImplementation
-internal data class SyncStreamState(
-    var targetCheckpoint: Checkpoint?,
-    var validatedCheckpoint: Checkpoint?,
-    var appliedCheckpoint: Checkpoint?,
-    var bucketSet: MutableSet<String>?,
-    var abortIteration: Boolean = false,
-)
 
 private class PendingCrudUpload(
     val done: CompletableDeferred<Unit>,
