--- common/src/commonMain/kotlin/com/powersync/sync/SyncDataBatch.kt
@@ -1,9 +1,0 @@
-package com.powersync.sync
-
-import kotlinx.serialization.Serializable
-
-@LegacySyncImplementation
-@Serializable
-internal data class SyncDataBatch(
-    val buckets: List<SyncLine.SyncDataBucket>,
-)
