+++ common/src/commonTest/kotlin/com/powersync/testutils/MockSyncService.kt
@@ -2,8 +2,6 @@ package com.powersync.testutils
 
 import app.cash.turbine.ReceiveTurbine
 import com.powersync.bucket.WriteCheckpointResponse
-import com.powersync.sync.LegacySyncImplementation
-import com.powersync.sync.SyncLine
 import com.powersync.sync.SyncStatusData
 import com.powersync.utils.JsonUtil
 import io.ktor.client.engine.HttpClientEngineBase
@@ -36,7 +34,6 @@ import kotlinx.serialization.json.JsonElement
  * function which makes it very hard to cancel the channel when the sync client closes the request stream. That is
  * precisely what we may want to test though.
  */
-@OptIn(LegacySyncImplementation::class)
 internal class MockSyncService(
     private val lines: () -> ReceiveChannel<Any>,
     private val syncLinesContentType: () -> ContentType,
